refactor: Move SubCommandContext & SubCommandContextBuilder into external file.
